Fixed MSVC warnings/errors

+++ include/json/Gason.hpp
@@ -115,7 +115,7 @@ namespace tson
                     m_path = path.parent_path();
 
                     file.seekg(0, std::ios::end);
-                    str.reserve(file.tellg());
+                    str.reserve(static_cast<size_t>(file.tellg()));
                     file.seekg(0, std::ios::beg);
 
                     str.assign((std::istreambuf_iterator<char>(file)),
